CDPD-3881 : Build changes to go with spark's hive-exec update.
The spark acid reader used to pull in some classes from spark's fork of
hive-exec.  However, as we're updating spark to use the standard cdpd
version of hive-exec, instead of its own fork, we need to update those
rules.  So we update imports slightly as we're now using the shaded
artifacts.

Change-Id: I7871a6c66fee49b14edef380146cb63d9775e822